+++ eggs/Nougat/src/main/java/com/android_n/egg/neko/NekoLand.java
@@ -253,7 +253,7 @@ public class NekoLand extends Activity implements PrefsListener {
                 public void onClick(View v) {
                     setContextGroupVisible(holder, false);
                     Cat cat = mCats[holder.getBindingAdapterPosition()];
-                    if (!ShareCatUtils.isRequireStoragePermissions()) {
+                    if (ShareCatUtils.isNotRequireStoragePermissions()) {
                         shareCat(cat);
                         return;
                     }