+++ feature/cat-editor/src/main/java/com/dede/android_eggs/cat_editor/CatEditor.kt
@@ -2,6 +2,7 @@
 
 package com.dede.android_eggs.cat_editor
 
+import android.graphics.Bitmap
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
@@ -22,21 +23,23 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Matrix
 import androidx.compose.ui.graphics.drawscope.withTransform
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.dede.android_eggs.cat_editor.CaptureControllerDelegate.Companion.rememberCaptureControllerDelegate
 import com.dede.android_eggs.cat_editor.CatParts.VIEW_PORT_SIZE
+import com.dede.android_eggs.cat_editor.Utilities.asAndroidMatrix
 import com.dede.android_eggs.cat_editor.Utilities.toInvert
 import dev.shreyaspatil.capturable.capturable
 import kotlin.math.max
@@ -45,8 +48,8 @@ import com.dede.android_eggs.resources.R as StringR
 
 private const val TAG = "CatEditor"
 
-private const val S_MIN = 0.5f
-private const val S_MAX = 3f
+private const val S_MIN = 0.3f
+private const val S_MAX = 5f
 
 private fun range(float: Float, max: Float, min: Float): Float {
     return min(max, max(float, min))
@@ -132,7 +135,7 @@ internal fun CatEditor(
             val infiniteTransition = rememberInfiniteTransition(label = "CatEditor_SelectedPart")
             val blendRatio by infiniteTransition.animateFloat(
                 initialValue = 0f,
-                targetValue = 0.5f,
+                targetValue = if (selectedPart != -1) 0.5f else 0f,
                 animationSpec = infiniteRepeatable(
                     animation = tween(700, easing = LinearEasing),
                     repeatMode = RepeatMode.Reverse,
@@ -140,19 +143,18 @@ internal fun CatEditor(
                 label = "HighlightColorBlend"
             )
 
-            val canvasMatrix = remember { Matrix() }
+            var canvasSize by remember { mutableStateOf(Size.Zero) }
+            val canvasMatrix = remember(canvasSize) { createCanvasMatrix(canvasSize) }
+
+            val bitmap: Bitmap? = remember(canvasSize) {
+                if (needAndroidCanvasDraw) createAndroidBitmap(canvasSize) else null
+            }
             Canvas(
                 contentDescription = stringResource(StringR.string.cat_editor),
                 modifier = Modifier
                     .fillMaxSize()
                     .aspectRatio(1f)
                     .capturable(captureController.getDelegate())
-                    .onSizeChanged {
-                        val size = min(it.width, it.height)
-                        canvasMatrix.reset()
-                        canvasMatrix.translate((it.width - size) / 2f, (it.height - size) / 2f)
-                        canvasMatrix.scale(size / VIEW_PORT_SIZE, size / VIEW_PORT_SIZE)
-                    }
                     .pointerInput(controllerImpl.isSelectEnabled) {
                         if (!controllerImpl.isSelectEnabled) {
                             return@pointerInput
@@ -180,6 +182,25 @@ internal fun CatEditor(
                         }
                     },
                 onDraw = {
+                    if (size != canvasSize) {
+                        // canvas size changed
+                        canvasSize = size
+                        return@Canvas
+                    }
+
+                    // native canvas draw
+                    if (needAndroidCanvasDraw && bitmap != null) {
+                        androidCanvasDraw(
+                            canvasMatrix.asAndroidMatrix(),
+                            bitmap,
+                            controllerImpl.colorList,
+                            selectedPart,
+                            blendRatio
+                        )
+                        return@Canvas
+                    }
+
+                    // compose canvas draw
                     withTransform({ transform(canvasMatrix) }) {
                         CatParts.drawOrders.forEachIndexed { index, pathDraw ->
                             var color = controllerImpl.colorList[index]
@@ -194,4 +215,12 @@ internal fun CatEditor(
             )
         }
     }
+}
+
+private fun createCanvasMatrix(size: Size): Matrix {
+    val matrix = Matrix()
+    val minDimension = size.minDimension
+    matrix.translate((size.width - minDimension) / 2f, (size.height - minDimension) / 2f)
+    matrix.scale(minDimension / VIEW_PORT_SIZE, minDimension / VIEW_PORT_SIZE)
+    return matrix
 }